refactor : user domain 파일 분리, 변수명 수정
