+++ src/user/type/error.ts
@@ -1,0 +1,6 @@
+const UserError = {
+	NOT_FOUND: 'USER_NOT_FOUND',
+	NOT_AUTHORIZED: 'NOT_AUTHORIZED',
+};
+
+export default UserError;
